requirements.txt
Flask
boto3
Pillow
requests
gunicorn
